+++ src/PortalBase/Builder/Support/TokenPriorityTrait.php
@@ -1,0 +1,20 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Heptacom\HeptaConnect\Portal\Base\Builder\Support;
+
+trait TokenPriorityTrait
+{
+    private ?int $priority = null;
+
+    public function getPriority(): ?int
+    {
+        return $this->priority;
+    }
+
+    public function setPriority(?int $priority): void
+    {
+        $this->priority = $priority;
+    }
+}