+++ src/PortalBase/Builder/Token/EmitterToken.php
@@ -5,9 +5,12 @@ declare(strict_types=1);
 namespace Heptacom\HeptaConnect\Portal\Base\Builder\Token;
 
 use Heptacom\HeptaConnect\Dataset\Base\Contract\DatasetEntityContract;
+use Heptacom\HeptaConnect\Portal\Base\Builder\Support\TokenPriorityTrait;
 
 class EmitterToken
 {
+    use TokenPriorityTrait;
+
     /**
      * @var class-string<DatasetEntityContract>
      */