+++ src/PortalBase/CHANGELOG.md
@@ -9,6 +9,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Add method `\Heptacom\HeptaConnect\Portal\Base\Builder\Builder\EmitterBuilder::priority` to sort flow component within the stack
+- Add method `\Heptacom\HeptaConnect\Portal\Base\Builder\Builder\ExplorerBuilder::priority` to sort flow component within the stack
+- Add method `\Heptacom\HeptaConnect\Portal\Base\Builder\Builder\HttpHandlerBuilder::priority` to sort flow component within the stack
+- Add method `\Heptacom\HeptaConnect\Portal\Base\Builder\Builder\ReceiverBuilder::priority` to sort flow component within the stack
+- Add method `\Heptacom\HeptaConnect\Portal\Base\Builder\Builder\StatusReporterBuilder::priority` to sort flow component within the stack
+- Add method `\Heptacom\HeptaConnect\Portal\Base\Builder\Token\EmitterToken::getPriority` and `\Heptacom\HeptaConnect\Portal\Base\Builder\Token\EmitterToken::setPriority` to sort flow component within the stack
+- Add method `\Heptacom\HeptaConnect\Portal\Base\Builder\Token\ExplorerToken::getPriority` and `\Heptacom\HeptaConnect\Portal\Base\Builder\Token\ExplorerToken::setPriority` to sort flow component within the stack
+- Add method `\Heptacom\HeptaConnect\Portal\Base\Builder\Token\HttpHandlerToken::getPriority` and `\Heptacom\HeptaConnect\Portal\Base\Builder\Token\HttpHandlerToken::setPriority` to sort flow component within the stack
+- Add method `\Heptacom\HeptaConnect\Portal\Base\Builder\Token\ReceiverToken::getPriority` and `\Heptacom\HeptaConnect\Portal\Base\Builder\Token\ReceiverToken::setPriority` to sort flow component within the stack
+- Add method `\Heptacom\HeptaConnect\Portal\Base\Builder\Token\StatusReporterToken::getPriority` and `\Heptacom\HeptaConnect\Portal\Base\Builder\Token\StatusReporterToken::setPriority` to sort flow component within the stack
+- Add method `\Heptacom\HeptaConnect\Portal\Base\Builder\FlowComponent::setDefaultPriority` to set default position for flow components within the stack per source package
+- Add method `\Heptacom\HeptaConnect\Portal\Base\Portal\Contract\PackageContract::getDefaultFlowComponentPriority` to set default position for flow components within the stack per source package
+
 ### Changed
 
 ### Deprecated